requirements.txt
psutil
rich
watchdog
yara-python
